// frontend/src/component/Button/DeviceButton/interface.d.ts
export interface IDeviceButtonProps {
    size: "small" | "big"
    type: "toggle" | "radio"
    isClicked?: boolean
}

export interface IWrapProps {
    size: "small" | "big"
    isClicked?: boolean
}

export interface IContainerProps {
    size: "small" | "big"
}

export interface IConWrapProps {
    size: "small" | "big"
}